fix: resolve two similar functions
