+++ home/urls.py
@@ -4,4 +4,5 @@ from . import views
 
 urlpatterns = [
     path("", views.index, name="index"),
+    path("search_view", views.search_view, name="search_view"),
 ]
