+++ questions/models.py
@@ -3,7 +3,6 @@ from django.core.validators import MinLengthValidator
 from django.db import models
 from taggit.managers import TaggableManager
 
-
 from lib.models import SoftDeleteModel
