--- search/templates/search/search_form.html
@@ -1,15 +1,0 @@
-{% extends "layouts/base.html" %}
-
-{% block content %}
-<div>
-    {% if results %}
-        <ul>
-            {% for result in results %}
-                <li>{{ result.title }}</li>
-            {% endfor %}
-        </ul>
-    {% else %}
-        <p>No results found.</p>
-    {% endif %}
-</div>
-{% endblock content %}
