--- search/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-
-from .views import search_view
-
-urlpatterns = [
-    path("", search_view, name="search_view"),
-]
